% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_line_fsa_list}
\alias{cell_line_fsa_list}
\title{A list of fsa files}
\format{
\subsection{\code{cell_line_fsa_list}}{

A list with 92 elements, each one being the contents of an fsa file:
}
}
\source{
\url{https://doi.org/10.1038/s41467-024-47485-0}
}
\usage{
cell_line_fsa_list
}
\description{
A list of fsa files read into R using trace::read_fsa() that for example data
}
\keyword{datasets}
