% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.pave.R
\name{compare_pave_adduct_distribution}
\alias{compare_pave_adduct_distribution}
\title{Adduct type counts for PAVE and TRACE barplots}
\usage{
compare_pave_adduct_distribution(adduct.df)
}
\arguments{
\item{adduct.df}{Output of [compare_pave_adduct_data()].}
}
\value{
A data frame with columns `source`, `adduct`, and `n`.
}
\description{
Adduct type counts for PAVE and TRACE barplots
}
