% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRACE_global_assignment.R
\name{.trace_global_assign_component}
\alias{.trace_global_assign_component}
\title{Global assignment within one connectivity component}
\usage{
.trace_global_assign_component(
  ig,
  nodes,
  blocked_chemforms,
  ppm.dyn,
  rt.tol.dyn,
  max_path_length = 6L,
  connection_cutoff = 0.5
)
}
\description{
Global assignment within one connectivity component
}
\keyword{internal}
