% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRACE_workflow.R
\name{.trace_init_temp}
\alias{.trace_init_temp}
\title{Initialize TRACE_temp container on MSdev object}
\usage{
.trace_init_temp(object)
}
\description{
Initialize TRACE_temp container on MSdev object
}
\keyword{internal}
