% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRACE_global_assignment.R
\name{.trace_retype_seed_edges}
\alias{.trace_retype_seed_edges}
\title{Per-edge CN consistency retyping (seed subgraph)}
\usage{
.trace_retype_seed_edges(cn.seed.net, cn.seed.formula)
}
\description{
Per-edge CN consistency retyping (seed subgraph)
}
\keyword{internal}
