% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRACE_global_assignment.R
\name{.trace_run_global_assignment}
\alias{.trace_run_global_assignment}
\title{Run PAVE-style global assignment for all CN seeds}
\usage{
.trace_run_global_assignment(
  cn.seed.ig.full,
  cn.seed,
  i.pol,
  object,
  pol,
  max_path_length = 6L,
  connection_cutoff = 0.5
)
}
\description{
Run PAVE-style global assignment for all CN seeds
}
\keyword{internal}
