% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{example_data}
\format{
\subsection{\code{example_data}}{

A genemapper output dataframe
}
}
\source{
\url{https://www.nature.com/articles/s41467-024-47485-0}
}
\usage{
example_data
}
\description{
example_data is genemapper output peak table
}
\keyword{datasets}
