% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_genemapper_alleles}
\alias{example_data_genemapper_alleles}
\title{example_data_genemapper_alleles}
\format{
\subsection{\code{example_data_genemapper_alleles}}{

A genemapper output dataframe
}
}
\source{
\url{https://www.nature.com/articles/s41467-024-47485-0}
}
\usage{
example_data_genemapper_alleles
}
\description{
example_data_genemapper_alleles is genemapper output peak table with genemapper
alleles called
}
\keyword{datasets}
