% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_repeat_table}
\alias{example_data_repeat_table}
\title{example_data_repeat_table}
\format{
\subsection{\code{example_data_repeat_table}}{

A genemapper output dataframe
}
}
\source{
\doi{10.1038/s41467-024-47485-0}
}
\usage{
example_data_repeat_table
}
\description{
example_data_repeat_table is data with repeats called
}
\keyword{datasets}
