% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{extract_trace_table}
\alias{extract_trace_table}
\title{Extract traces}
\usage{
extract_trace_table(fragments_list)
}
\arguments{
\item{fragments_list}{a list of fragments objects}
}
\value{
A dataframe of the raw trace data. Each row representing a single scan.
}
\description{
Extract the raw trace from a list of fragments objects
}
\examples{
fsa_list <- lapply(cell_line_fsa_list, function(x) x$clone())
# import data with read_fsa() to generate an equivalent list to cell_line_fsa_list
test_fragments <- trace(fsa_list, grouped = TRUE, metadata_data.frame = metadata)

extracted_traces <- extract_trace_table(test_fragments)

}
