% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{generate_trace_template}
\alias{generate_trace_template}
\title{Generate a Quarto file that has the instability pipeline preset}
\usage{
generate_trace_template(
  file_name = NULL,
  correction = "batch",
  samples_grouped = TRUE
)
}
\arguments{
\item{file_name}{Name of file to create}

\item{correction}{select either "none", "batch" or "repeat" to indicate if the functionality for correcting repeat size using size standards across batches be included in the pipeline. See \code{\link[=add_metadata]{add_metadata()}} & \code{\link[=call_repeats]{call_repeats()}} for more info.}

\item{samples_grouped}{Indicates if the functionality for grouping samples for metrics calculations should be included in the pipeline. See \code{\link[=add_metadata]{add_metadata()}} & \code{\link[=assign_index_peaks]{assign_index_peaks()}} for more info.}
}
\value{
A Quarto template file for repeat instability analysis
}
\description{
Generate a Quarto file that has the instability pipeline preset
}
\examples{

if (interactive()) {
   generate_trace_template("test")
}


}
