% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{metadata}
\format{
\subsection{\code{metadata}}{

A genemapper output dataframe
}
}
\source{
\doi{10.1038/s41467-024-47485-0}
}
\usage{
metadata
}
\description{
This is a dataframe containing the metadata information for the example data
}
\keyword{datasets}
