% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_data_channels}
\alias{plot_data_channels}
\title{plot_data_channels}
\usage{
plot_data_channels(fragments_list, sample_subset = NULL, n_facet_col = 1)
}
\arguments{
\item{fragments_list}{A list of fragments objects.}

\item{sample_subset}{A character vector of unique ids for a subset of samples to plot}

\item{n_facet_col}{A numeric value indicating the number of columns for faceting in the plot.}
}
\value{
a plot of the raw data channels
}
\description{
Plot the raw data from the fsa file
}
\details{
A plot of the raw data channels in the fsa file.

These plots are made using base R plotting. Sometimes these fail to render in the viewing panes of IDEs (eg you get the error 'Error in \code{plot.new()}: figure margins too large)'. If this happens, try saving the plot as a pdf using traditional approaches (see grDevices::pdf). To get it to render in the IDE pane, trying matching \code{n_facet_col} to the number of samples you're attempting to plot, or using \code{sample_subset} to limit it to a single sample.
}
\examples{

plot_data_channels(cell_line_fsa_list[1])

}
