% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_chem.R
\name{get_PAVE_from_MSdev}
\alias{get_PAVE_from_MSdev}
\title{PAVE analysis}
\usage{
get_PAVE_from_MSdev(object)
}
\arguments{
\item{object}{MSdev object}
}
\value{
PAVE
}
\description{
PAVE from MSdev.
}
\section{Functions}{
\itemize{
\item \code{get_PAVE_from_MSdev()}: initialize from MSdev

}}
\examples{
get_PAVE_from_MSdev(msdev)
}
