% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace-package.R
\docType{package}
\name{trace-package}
\alias{trace-package}
\title{trace: Tandem Repeat Analysis by Capillary Electrophoresis}
\description{
A pipeline for short tandem repeat instability analysis from fragment analysis data. Inputs of fsa files or peak tables, and a user supplied metadata data-frame. The package identifies ladders, calls peaks, identifies the modal peaks, calls repeats, then calculates repeat instability metrics (e.g. expansion index from Lee et al. (2010) \doi{10.1186/1752-0509-4-29}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://zachariahmclean.github.io/trace/}
}

}
\author{
\strong{Maintainer}: Zachariah McLean \email{zachariah.louis.mclean@gmail.com} (\href{https://orcid.org/0000-0002-0968-0538}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kevin Correia
}

Other contributors:
\itemize{
  \item Andrew Jiang [contributor]
}

}
\keyword{internal}
