% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRACE_workflow.R
\name{TRACE_dynamic_filter}
\alias{TRACE_dynamic_filter}
\title{Apply dynamic m/z and RT filtering for TRACE network}
\usage{
TRACE_dynamic_filter(object, i.pol)
}
\arguments{
\item{object}{MSdev object.}

\item{i.pol}{Polarity index. `0` for negative and `1` for positive.}
}
\value{
Updated MSdev object with dynamic filter outputs in
  `object@advancedAna$TRACE_temp[[pol]]`.
}
\description{
Apply dynamic m/z and RT filtering for TRACE network
}
