% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRACE_workflow.R
\name{TRACE_export}
\alias{TRACE_export}
\title{Export TRACE results to xlsx}
\usage{
TRACE_export(object, file = NULL)
}
\arguments{
\item{object}{MSdev object with TRACE results in `object@advancedAna$TRACE`.}

\item{file}{Output xlsx path. Default is
`file.path(object@projectInfo$projectDir, "Statistic", "TRACE.xlsx")`
when `projectDir` is set; otherwise
`paste0(object@projectInfo$MSdevFile, ".TRACE.xlsx")`.}
}
\value{
Invisibly, the output file path.
}
\description{
Writes final TRACE annotation tables from `object@advancedAna$TRACE` to an
Excel workbook. When available, CN labelling-ratio tables from
`object@advancedAna$TRACE_temp` are included as additional sheets.
}
