% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRACE_workflow.R
\name{TRACE_get_CN_net}
\alias{TRACE_get_CN_net}
\title{Build initial TRACE CN/adduct/isotope/fragment candidate networks}
\usage{
TRACE_get_CN_net(object, i.pol, rt.tol = 10, ppm = 5)
}
\arguments{
\item{object}{MSdev object.}

\item{i.pol}{Polarity index. `0` for negative and `1` for positive.}

\item{rt.tol}{RT tolerance in seconds.}

\item{ppm}{m/z tolerance in ppm.}
}
\value{
Updated MSdev object with intermediate data in
  `object@advancedAna$TRACE_temp[[pol]]`.
}
\description{
Build initial TRACE CN/adduct/isotope/fragment candidate networks
}
