% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRACE_workflow.R
\name{TRACE_network_assignment}
\alias{TRACE_network_assignment}
\title{Build TRACE seed network assignment}
\usage{
TRACE_network_assignment(object, i.pol)
}
\arguments{
\item{object}{MSdev object.}

\item{i.pol}{Polarity index. `0` for negative and `1` for positive.}
}
\value{
Updated MSdev object with seed network stored in
  `object@advancedAna$TRACE_temp[[pol]]$cn.seed.ig`.
}
\description{
Build TRACE seed network assignment
}
