% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRACE_workflow.R
\name{TRACE_workflow}
\alias{TRACE_workflow}
\title{Run the refactored TRACE workflow}
\usage{
TRACE_workflow(
  object,
  rt.tol = 10,
  ppm = 5,
  cpdb = "d:/data/2025.12.26.TRACE2/trace.cp.db.xlsx"
)
}
\arguments{
\item{object}{MSdev object.}

\item{rt.tol}{RT tolerance in seconds.}

\item{ppm}{m/z tolerance in ppm.}

\item{cpdb}{Path to compound table xlsx used for candidate assignment.}
}
\value{
Updated MSdev object with final outputs in `object@advancedAna$TRACE`
  and intermediate outputs in `object@advancedAna$TRACE_temp`.
}
\description{
Run the refactored TRACE workflow
}
