% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add Metadata to Fragments List}
\usage{
add_metadata(
  fragments_list,
  metadata_data.frame,
  unique_id = "unique_id",
  metrics_group_id = "metrics_group_id",
  metrics_baseline_control = "metrics_baseline_control",
  batch_run_id = "batch_run_id",
  batch_sample_id = "batch_sample_id"
)
}
\arguments{
\item{fragments_list}{A list of fragment objects to which metadata will be added.}

\item{metadata_data.frame}{A data frame containing the metadata information.}

\item{unique_id}{(required) A character string indicating the column name for unique sample identifiers in the metadata.}

\item{metrics_group_id}{(optional) A character string indicating the column name for sample group identifiers in the metadata. To skip, provide NA.}

\item{metrics_baseline_control}{(optional) A character string indicating the column name for baseline control indicators in the metadata. To skip, provide NA.}

\item{batch_run_id}{(optional) A character string indicating the column name for plate identifiers in the metadata.  To skip, provide NA.}

\item{batch_sample_id}{(optional) A character string indicating the column name for an id of the size standard. For example, a sample code. This can then be used to correct batch effects across different fragment analysis runs. To skip, provide NA.}
}
\value{
A modified list of fragment objects with added metadata.
}
\description{
This function adds metadata information to a list of fragments.
}
\details{
This function adds specified metadata attributes to each fragment in the list.
It matches the unique sample identifiers from the fragments list with those in the metadata data frame.
To skip any of the optional columns, make parameter NA.
}
\examples{

gm_raw <- trace::example_data
metadata <- trace::metadata

test_fragments <- peak_table_to_fragments(gm_raw,
  data_format = "genemapper5",
  dye_channel = "B",
  min_size_bp = 300
)

add_metadata(
  fragments_list = test_fragments,
  metadata_data.frame = metadata,
  unique_id = "unique_id",
  metrics_group_id = "metrics_group_id",
  metrics_baseline_control = "metrics_baseline_control",
  batch_run_id = "batch_run_id",
  batch_sample_id = "batch_sample_id"
)

# skip unwanted metadata by using NA

add_metadata(
  fragments_list = test_fragments,
  metadata_data.frame = metadata,
  unique_id = "unique_id",
  metrics_group_id = "metrics_group_id",
  metrics_baseline_control = "metrics_baseline_control",
  batch_run_id = NA,
  batch_sample_id = NA
)

}
