% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_index_peaks.R
\name{assign_index_peaks}
\alias{assign_index_peaks}
\title{Assign index peaks}
\usage{
assign_index_peaks(
  fragments_list,
  grouped = FALSE,
  index_override_dataframe = NULL
)
}
\arguments{
\item{fragments_list}{A list of "fragments_repeats" class objects representing
fragment data.}

\item{grouped}{Logical value indicating whether samples should be grouped to
share a common index peak. \code{FALSE} will assign the sample's own modal allele as the index peak. \code{TRUE} will use metadata to assign the index peak based on the modal peak of another sample (see below for more details).}

\item{index_override_dataframe}{A data.frame to manually set index peaks.
Column 1: unique sample IDs, Column 2: desired index peaks (the order of the
columns is important since the information is pulled by column position rather
than column name). Closest peak in each sample is selected so the number needs to just be approximate.}
}
\value{
A list of \code{"fragments_repeats"} objects with index_repeat and index_height added.
}
\description{
Assign index peaks in preparation for calculation of instability metrics
}
\details{
A key part of several instability metrics is the index peak. This is the repeat
length used as the reference peak for relative instability metrics calculations, like expansion index or average repeat gain.
This is usually the the inherited repeat length of a mouse, or the modal repeat length for the cell line at a starting time point.

If \code{grouped} is set to \code{TRUE}, this function groups the samples by their \code{metrics_group_id} and uses the samples set as \code{metrics_baseline_control} to set the index peak. Use \code{\link[=add_metadata]{add_metadata()}} to set these variables. This is useful for cases like inferring repeat size of inherited alleles from mouse tail data. If the samples that are going to be used to assign index peak are from different fragment analysis runs, use \code{batch_correction} in \code{\link[=call_repeats]{call_repeats()}} to make sure the systematic differences between runs are corrected and the correct index peak is assigned.

For mice, if just a few samples have the inherited repeat height shorter than the expanded population, you could not worry about this and instead use the \code{index_override_dataframe}. This can be used to manually override these assigned index repeat values (irrespective of whether \code{grouped} is TRUE or FALSE).

As a final option, the index peak could be manually assigned directly to a \link{fragments_repeats} class using the internal setter function \code{\link[=fragments_repeats$set_index_peak]{fragments_repeats$set_index_peak()}}.
}
\examples{


file_list <- trace::cell_line_fsa_list

ladder_list <- find_ladders(file_list)

fragments_list <- find_fragments(ladder_list,
  min_bp_size = 300
)

allele_list <- find_alleles(
  fragments_list = fragments_list
)
repeats_list <- call_repeats(
  fragments_list = allele_list
)

metadata_added_list <- add_metadata(
  fragments_list = repeats_list,
  metadata_data.frame = trace::metadata
)

index_assigned <- assign_index_peaks(metadata_added_list,
                                    grouped = TRUE)

plot_traces(index_assigned[1], xlim = c(100,150))






}
