% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{extract_alleles}
\alias{extract_alleles}
\title{Extract Modal Peaks}
\usage{
extract_alleles(fragments_list)
}
\arguments{
\item{fragments_list}{A list of fragments objects containing fragment data.}
}
\value{
A dataframe containing modal peak information for each sample
}
\description{
Extracts modal peak information from each sample in a list of fragments.
}
\examples{
fsa_list <- lapply(cell_line_fsa_list, function(x) x$clone())
# import data with read_fsa() to generate an equivalent list to cell_line_fsa_list
test_fragments <- trace_main(fsa_list, grouped = TRUE, metadata_data.frame = metadata)

extract_alleles(test_fragments)

}
