% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{extract_fragments}
\alias{extract_fragments}
\title{Extract All Fragments}
\usage{
extract_fragments(fragments_list)
}
\arguments{
\item{fragments_list}{A list of fragments objects containing fragment data.}
}
\value{
A dataframe containing peak data for each sample
}
\description{
Extracts peak data from each sample in a list of fragments.
}
\examples{
fsa_list <- lapply(cell_line_fsa_list, function(x) x$clone())
# import data with read_fsa() to generate an equivalent list to cell_line_fsa_list
test_fragments <- trace_main(fsa_list, grouped = TRUE, metadata_data.frame = metadata)

extract_fragments(test_fragments)

}
