% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{extract_trace_table}
\alias{extract_trace_table}
\title{Extract traces}
\usage{
extract_trace_table(fragments_list)
}
\arguments{
\item{fragments_list}{a list of fragments objects}
}
\value{
A dataframe of the raw trace data. Each row representing a single scan.
}
\description{
Extract the raw trace from a list of fragments objects
}
\examples{
fsa_list <- lapply(cell_line_fsa_list[1], function(x) x$clone())

find_ladders(fsa_list, show_progress_bar = FALSE)

extracted_traces <- extract_trace_table(fsa_list)

}
