% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_alleles.R
\name{find_alleles}
\alias{find_alleles}
\title{Find Alleles}
\usage{
find_alleles(
  fragments_list,
  peak_region_size_gap_threshold = 6,
  peak_region_height_threshold_multiplier = 1
)
}
\arguments{
\item{fragments_list}{A list of fragment objects containing peak data.}

\item{peak_region_size_gap_threshold}{Gap threshold for identifying peak regions. The peak_region_size_gap_threshold is a parameter used to determine the maximum allowed gap between peak sizes within a peak region. Adjusting this parameter affects the size range of peaks that can be grouped together in a region. A smaller value makes it more stringent, while a larger value groups peaks with greater size differences, leading to broader peak regions that may encompass wider size ranges.}

\item{peak_region_height_threshold_multiplier}{Multiplier for the peak height threshold. The peak_region_height_threshold_multiplier parameter allows adjusting the threshold for identifying peak regions based on peak heights. Increasing this multiplier value will result in higher thresholds, making it more stringent to consider peaks as part of a peak region. Conversely, reducing the multiplier value will make the criteria less strict, potentially leading to more peaks being grouped into peak regions. It's important to note that this parameter's optimal value depends on the characteristics of the data and the specific analysis goals. Choosing an appropriate value for this parameter can help in accurately identifying meaningful peak regions in the data.}
}
\value{
A list of fragments with identified main alleles.
}
\description{
This function identifies main allele within each fragment object.
}
\details{
This function finds the main alleles for each fragment in the list by identifying clusters of peaks ("peak regions") with the highest signal intensities. This is based on the idea that PCR amplicons of repeats have clusters of peaks (from somatic mosaicism and PCR artifacts) that help differentiate the main allele of interest from capillary electrophoresis noise/contamination.

The tallest of peaks will be selected as the allele. This means that if your sample has multiple alleles, you'll need to make sure that your data is subsetted to exclude that second allele. For example in a human cell line with a shorter repeat, we recommend using \code{min_bp_size} in \code{\link[=find_fragments]{find_fragments()}} to make sure that the smaller allele is excluded. Otherwise, the smaller taller allele will be set as the allele.

The parameters \code{peak_region_height_threshold_multiplier} and \code{peak_region_size_gap_threshold} will only need to be adjusted if peaks are not being found for some reason. They influence the criteria for identifying peak regions, and finding the right balance between them is crucial.
}
\examples{
file_list <- trace::cell_line_fsa_list

test_ladders <- find_ladders(file_list[1])

fragments_list <- find_fragments(test_ladders,
  min_bp_size = 300
)


test_alleles <- find_alleles(
  fragments_list = fragments_list,
  peak_region_size_gap_threshold = 6,
  peak_region_height_threshold_multiplier = 1
)
}
