% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ladders.R
\name{find_ladders}
\alias{find_ladders}
\title{Ladder and bp sizing}
\usage{
find_ladders(
  fragments_trace,
  ladder_sizes = c(50, 75, 100, 139, 150, 160, 200, 250, 300, 340, 350, 400, 450, 490,
    500),
  spike_location = NULL,
  minimum_peak_signal = NULL,
  zero_floor = FALSE,
  scan_subset = NULL,
  ladder_selection_window = 5,
  max_combinations = 2500000,
  smoothing_window = 21,
  show_progress_bar = TRUE
)
}
\arguments{
\item{fragments_trace}{list from 'read_fsa' function}

\item{ladder_sizes}{numeric vector: bp sizes of ladder used in fragment analysis.
defaults to GeneScan™ 500 LIZ™}

\item{spike_location}{numeric: indicate the scan number to start looking for
ladder peaks. Usually this can be automatically found (when set to NULL) since
there's a big spike right at the start. However, if your ladder peaks
are taller than the big spike, you will need to set this starting scan
number manually.}

\item{minimum_peak_signal}{numeric: minimum height of peak from smoothed signal.}

\item{zero_floor}{logical: if set to TRUE, all negative values will be set to zero.
This can help deal with cases where there are peaks in the negative direction
that interfere with peak detection.}

\item{scan_subset}{numeric vector (length 2): filter the ladder and data signal
between the selected scans (eg scan_subset = c(3000, 5000)).
to pracma::savgol().}

\item{ladder_selection_window}{numeric: in the ladder assigning algorithm,
the we iterate through the scans in blocks and test their linear fit ( We can assume that the ladder is linear over a short distance)
This value defines how large that block of peaks should be.}

\item{max_combinations}{numeric: what is the maximum number of ladder
combinations that should be tested}

\item{smoothing_window}{numeric: ladder signal smoothing window size passed}

\item{show_progress_bar}{show progress bar}
}
\value{
list of fragments_trace objects
}
\description{
Find the ladder peaks in and use that to call bp size
}
\details{
This function takes a list of fragments_trace files (the output from read_fsa) and identifies
the ladders in the ladder channel which is used to call the bp size. The output
is a list of fragments_traces. bp sizes are assigned using the local Southern
method. Basically, for each data point, linear models are made for the lower
and upper 3 size standard and the predicted sizes are averaged.

The ladder peaks are assigned from largest to smallest. I would recommend excluding
size standard peaks less than 50 bp (eg size standard 35 bp).

Each ladder should be manually inspected to make sure that is has been correctly
assigned.
}
\examples{

file_list <- trace::cell_line_fsa_list

test_ladders <- find_ladders(file_list, show_progress_bar = FALSE)

# Manually inspect the ladders
plot_ladders(test_ladders[1])

}
\seealso{
\code{\link[=fix_ladders_auto]{fix_ladders_auto()}} and \code{\link[=fix_ladders_interactive]{fix_ladders_interactive()}} to fix ladders with
incorrectly assigned peaks. \code{\link[=plot_ladders]{plot_ladders()}} to plot the assigned ladder
peaks onto the raw ladder signal.
}
