% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ladders}
\alias{plot_ladders}
\title{Plot ladder}
\usage{
plot_ladders(
  fragments_list,
  n_facet_col = 1,
  sample_subset = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{fragments_list}{A list of fragments objects containing fragment data.}

\item{n_facet_col}{A numeric value indicating the number of columns for faceting in the plot.}

\item{sample_subset}{A character vector of unique ids for a subset of samples to plot}

\item{xlim}{the x limits of the plot. A numeric vector of length two.}

\item{ylim}{the y limits of the plot. A numeric vector of length two.}
}
\value{
a plot of ladders
}
\description{
Plot the ladder signal
}
\examples{

fsa_list <- lapply(cell_line_fsa_list, function(x) x$clone())
# import data with read_fsa() to generate an equivalent list to cell_line_fsa_list
fragments_list <- trace_main(fsa_list)

# Manually inspect the ladders
plot_ladders(fragments_list[1])

}
