% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_repeat_correction_model}
\alias{plot_repeat_correction_model}
\title{Plot Repeat Correction Model}
\usage{
plot_repeat_correction_model(fragments_list, batch_run_id_subset = NULL)
}
\arguments{
\item{fragments_list}{A list of fragments_repeats class objects obtained from the 'call_repeats' function when the 'repeat_length_correction' was either 'from_metadata' or 'from_genemapper'.}
}
\value{
A base R graphic object displaying the repeat correction model results.
}
\description{
Plots the results of the repeat correction model for a list of fragments.
}
\examples{


gm_raw <- instability::example_data
metadata <- instability::metadata

test_fragments <- peak_table_to_fragments(
  gm_raw,
  data_format = "genemapper5",
  dye_channel = "B"
)

test_alleles <- find_alleles(
  fragments_list = test_fragments,
  number_of_peaks_to_return = 2,
  peak_region_size_gap_threshold = 6,
  peak_region_height_threshold_multiplier = 1
)

test_metadata <- add_metadata(
  fragments_list = test_alleles,
  metadata_data.frame = metadata
)

test_repeats_corrected <- call_repeats(
  fragments_list = test_metadata,
  repeat_calling_algorithm = "none",
  assay_size_without_repeat = 87,
  repeat_size = 3,
  repeat_length_correction = "from_metadata"
)

plot_repeat_correction_model(test_repeats_corrected)

}
