% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fsa.R
\name{read_fsa}
\alias{read_fsa}
\title{Read fsa file}
\usage{
read_fsa(files, ladder_channel = "DATA.105", signal_channel = "DATA.1")
}
\arguments{
\item{files}{a chr vector of fsa file names.}

\item{ladder_channel}{string: which channel in the fsa file contains the
ladder signal}

\item{signal_channel}{string: which channel in the fsa file contains the data
signal}
}
\value{
A list of ABIF formatted files generated by seqinr::read.abif(). List
elements are named by the file name.
}
\description{
This function is just a wrapper of seqinr::read.abif
}
\examples{
# files <- list.files("example_directory/", full.names = TRUE, pattern = ".fsa")
# read_fsa(files)

}
\seealso{
\code{\link[seqinr:read.abif]{seqinr::read.abif()}}
}
