% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRACE_workflow.R
\name{TRACE_annotate}
\alias{TRACE_annotate}
\title{Annotate TRACE seed network candidates and finalize TRACE result}
\usage{
TRACE_annotate(
  object,
  i.pol,
  cpdb = "d:/data/2025.12.26.PAVE2/trace.cp.db.xlsx"
)
}
\arguments{
\item{object}{MSdev object.}

\item{i.pol}{Polarity index. `0` for negative and `1` for positive.}

\item{cpdb}{Path to compound table xlsx used for candidate assignment.}
}
\value{
Updated MSdev object with `object@advancedAna$TRACE[[pol]]`.
}
\description{
Annotate TRACE seed network candidates and finalize TRACE result
}
